Warning
git-svn-id: https://subversion.assembla.com/svn/fakesmc@370 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ Plug-ins/FakeSMCRadeon/ATICard.cpp
@@ -69,7 +69,7 @@ bool ATICard::initialize()
 			break;
 			
 		default:
-			InfoLog("sorry, but your card %04lx is not supported!\n", chipID>>16);
+			InfoLog("sorry, but your card %04lx is not supported!\n", (long unsigned int)chipID>>16);
 			return false;
 	}
 	return true;
